requirements.txt
langchain==0.0.123
python-dotenv==1.0.0
faiss-cpu==1.7.3
langchain==0.0.115
llama-index==0.4.31
openai==0.27.2
pandas==1.5.3
pdfminer==20191125
pdfminer.six==20221105
tokenizers==0.13.2
transformers==4.27.2
unstructured==0.5.6
unstructured-inference==0.2.11
